+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changelog
 - **Fix**: Fix qualifiers on bindings not getting seen by extended child graphs.
 - **Fix**: Fix qualifiers getting ignored on accessors from `@Includes` dependencies.
 - **Change**: `InstanceFactory` is no longer a value class. This wasn't going to offer much value in practice.
+- **Change**: Change debug reports dir to be per-compilation rather than per-platform.
 
 0.2.0
 -----